test : ChattingService 테스트 생성
